[storage] Add a wait method to ExecutionScope.
At the moment, it's not possible to implement a vfs server that can wait
for all connections to be closed, so this adds a wait message that can
be used.

Also changed vfs tests to run as a v2 test component rather than v1.

Bug: 53796
Change-Id: I02bd8dd85512adfd166f8f11ccf3ec9e46d080c4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/402722
Reviewed-by: Simon Shields <simonshields@google.com>
Testability-Review: Simon Shields <simonshields@google.com>
Commit-Queue: Chris Suter <csuter@google.com>