[blobfs] Implement VmoRead for new pager
Plumbs the VmoRead request for the new pager from Blob through to the
UserPager. Minor changes are made to UserPager and PageWatcher (the old
paging infrastructure) to allow this to happen in the same way whether
the new pager is enabled or not.

Make VMO parameters const. The zx library sues const on a VMO to mean
the C++ object, not the udnerlying system object, is constant, while the
old pager library assume const meant both were constant. This mismatch
caused some problems when hooking up to the blobfs code.

The new pager code is not currently compiled by default and does not yet
work. The next step will be creating the paged VMO properly.

Change-Id: Id9a5848fa335d0204c1763da6546eedaf72072ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/497719
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>