[vfs] Deprecate and remove pcb file
- Replace pcb static/const read_only functions with vmo equivalents.
- Replace pcb::read_write with vmo::read_write
- Replace all but one pcb::write_only with vmo:: write_only

Below is is sum of memory used in bytes, with and without this patch, by
all the processes ~74 seconds after boot as seen in ps output.

        | Private   | Shared
--------+-----------+----------
Without | 625290854 | 620360499
With.   | 621666918 | 616597094

Memory consumed with this patch seems to be less than without patch
indicating noise and this patch doesn't have significant effect on
memory footprint.

Bug: 66267

Change-Id: Idbf549aa14879ab29a4fe9b0220fc1ab9b2a7282
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/493669
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Vikram Auradkar <auradkar@google.com>