[fxfs] Device, Buffer and BufferAllocator
Device is the interface for reading and writing from a block device.

Buffer is a sliceable, contiguous buffer which is taken out of a special
memory region (for example, from a mapped VMO) which is appropriate for
transferring data to and from a block device. Buffers are owned by the
code which requests them, and automatically clean themselves up when
going out of scope.

BufferAllocator is the interface for creating Buffers which can be used
to transfer data to and from a Device. Generally, clients will simply
ask the Device for a reference to its allocator, and allocate blocks out
of that.

Each Device implementation is expected to implement its own BufferSource
which is the backing pool for Buffers. For example, a Device which is
backed by a real block FIFO would allocate Buffer objects out of a VMO
which is attached to the block device (i.e. has a vmoid). A later CL
will add such a Device.

Change-Id: I5067082bc6767d128e2a2dec8c57f3181ed6c9d9
Tested: Unit.
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/500279
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>