[storage] Provide Rust client interface to get the FilesystemInfo
FilesystemInfo structure can be used to get the amount of free and used
space on a mounted file system.

To verify correctness, the existing minsfs and blobfs unit tests in
fs_management were updated to take a snapshot of FilesystemInfo and
compare it against another instance once data is written to the filesystem.

Test: - fx --dir=out/minfs_stress set core.x64 --with
	src/lib/storage/fs_management:tests
	- fx test fs-management-rust-tests
Bug: 70395
Change-Id: I9b9ef19d88b22e5c45651c9d5e232da7658abf16
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/486419
Commit-Queue: Kamen Vitanov <kamen@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>