[fit-promise] Add dependencies on fit-promise
This is a new SDK library into which the "promise"-related content of
libfit will be moved. This is the first step in a soft transition.
Currently fit-promise is empty; after everything that should be
depending on it is doing so, the content will be migrated from fit.

Bug: 57348
Change-Id: Id783da940b6258784ec288fbbbf7b71c7bc92bd0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/507119
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>