[block] Make VmoId interiorly mutable
Wraps VmoId's interior ID in an Atomic, so that it can be interiorly
mutated. This is necessary to "take" a VmoId when all we have is a
reference to it.

For example, when doing asynchronous cleanup of some object that owns a
VmoId, the object needs to unregister the VmoId with their block client.
That object would either need to be &mut self in the cleanup context
(which is an onerous requirement), or it would need to wrap VmoId in a
Mutex. All of this can be made easier and simpler by making VmoId
interiorly mutable.

Tested: Built.
Change-Id: I28a92a0afd3a61331c68c6d99109d074d73ed43d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/509704
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>