[fxfs] Fix Admin/Shutdown
To avoid races in tests that immediately try and open the block device
after shutting down a filesystem, we need to ensure we've closed the
block device before sending the response to Admin/Shutdown.

In order to facilitate that, I needed to add a close method to
BlockClient, and in the process of doing that, I realised that it wasn't
properly async (the FIDL proxy we were using was synchronous), so I
fixed that and that involved separating out the synchronous and
asynchronous APIs.  Unfortunately, it does mean that Cache cannot be
used with a fake device, but that should be OK; we had nothing that was
relying on that feature.

Change-Id: I10be180f735cc0642c8e3dfd5608b1c28c435c36
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525840
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>