[vfs] Separate out node registration from vmo.
Previously a PagedVnode being registered for paging requests was tied to
the existence of the paged vmo. But blobfs wants to cache nodes as the
associated vmo even when the file is closed. The reference owned by the
PagedVfs interferes with this by forcing the node to be alive.

This change separates out the existence of the paged vmo from whether it
is registered for paging requests. This allows blobfs to manage its
cache how it wants without interfering with the lifetimes of the blobs.

Tie the registration of the paged vmo to the HasReferences() state of
the blob. This now requires that all callers wishing to read, write, or
verify a blob must first open references. This mainly affects tests but
the verify and upgrade cases in blobfs also require changes. Some
helper classes are added to make managing the open state easier.

Remove some asserts from Blob::Purge and instead return the error. This
was getting hit in a migration error test when the newly-added Close()
call failed. This is expected to fail in some cases (in the test the
device is full) so should not be an assertion.

Ensure the Vnode (Blob) is freed outside of its stack frame in the new
pager when unregistering.

Bug: 51111
Change-Id: Ibff39b7a17e3273907041d3789e64f1e1713bc72
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/525446
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Martin Lindsay <mlindsay@google.com>