[fxfs] Add a writeback cache
Introduces a writeback cache to fxfs which buffers writes on files until
they are fsynced() or their last handle closes.

This also changes the read path for files to read from the cache when
possible. Reading an absent part of the file will load the surrounding
32KiB of data into the cache for later use.

The writeback cache was written in a way where we can (hopefully) easily
move it out of fxfs once its design settles. Its only fxfs dependency is
on the allocator::Reservation construct, which we'll need to move out to
the VFS (with appropriate callbacks/hooks back to fxfs).

Change-Id: I24b8cda28a9c5e79bb69a32b005ba517897b85d8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/547024
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>