[vfs][rust] Use slices for reading
We previously changed the vfs File interface to use storage_device
Buffer objects, which would permit zero copy IO from the VFS through to
the block device and back.

We've realized there are ergonomics issues with these Buffer objects in
the VFS, and that much of this can be simplified by returning to regular
slices. Moreover, we expect the IO paths to change drastically in the
near future, in the following stages:
- (Ongoing) reads/writes are copied into a writeback cache, with no
  alignment requirements; this is done in fxfs, the main rust filesystem
  we currently care about the performance of.
- (Near term) reads/writes could go through zx_vmo_read/write if we
  implement a pager, which simplifies the number of paths we maintain.
- (Longer term) reads/writes could be removed from fuchsia.io completely
  in favor of the zx::stream interface.

Fixed: 81847
Fixed: 82060
Change-Id: I4320d79fb9dd2770e90d801e28249bdd71fd809e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/568041
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>