[storage] Fixed racy link
The handling of link in the Rust VFS was racy. This CL adds a test and
fixes it.

Avoiding the races and keeping it working for pseudo directories is hard
because the source can be a number of different concrete types and locks
need to be maintained.  Support for hard links in mutable
pseudo-directories is not actually needed right now, so I've removed
support for that for now.  If the need arises, we can add back support
but it will involve a fair amount of restructuring.

Fixed: 81237
Change-Id: I50b3a0220627baddf75b9279832f237f25a8bff5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/581182
API-Review: Chris Suter <csuter@google.com>
Commit-Queue: Chris Suter <csuter@google.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>