[fshost] Add support for binding nand-broker to nand devices.
This CL extends the block watcher to also watch for NAND devices.
On Chromebook, a userspace component (//src/storage/flashmap) is
responsible for interacting with the NAND, and it uses the nand-broker
to do that. This CL eliminates the need for its clients to handle
binding the nand-broker manually.

Multiply: fshost-integration-test
Multiply: paver-test
Bug: 81685
Change-Id: I6d8ed0bf9450b67e4542dd01d890755067b53f61
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/582121
Reviewed-by: Abdulla Kamar <abdulla@google.com>
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>