[llcpp] Use generic transports in OutgoingMessage and IncomingMessage
This CL updates the interface of OutgoingMessage and IncomingMessage to
use generic transports rather than channels.

As a consequence, the API is updated so that channel-specific methods
are removed making it less convenient to use these APIs with channels
and requiring the use of an API in ::fidl::internal to read and write to
a channel. An alternative would be to leave the existing APIs on the
classes and route the calls through the generic APIs. However, this will
make less and less sense as more transports are added. Another
alternative would be to move the APIs to create transports from
::fidl::internal to ::fidl.

Bug: 85734
Change-Id: Ifca618af84574f85db2e006177719db604b6690f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/592405
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
API-Review: Ian McKellar <ianloic@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>