[vfs] Move QueryFilesystem to FuchsiaVfs.
This is designed to make it easier to implement QueryFilesystem on all
filesystems and to unify on one implementation. Currently this is
exposed in different ways.

With this change, a filesystem needs only to implement GetFilesystemInfo
once on their VFS and it will work for all nodes, and they need only to
instantiate route the fs::QueryService to implement that API using the
same implementation.

Adds more Vfs backpointers to the root nodes in the Blobfs root.

Change-Id: I9cb35b4c5ad3b8691eabba26b9d4d86d266875c6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/596749
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>