[llcpp] Remove transport_type field from C messages
This CL moves the transport_type field from the C message structs to the
LLCPP IncomingMessage and OutgoingMessage classes. The idea is to
require C messages to always use the FIDL_TRANSPORT_TYPE_CHANNEL type.

Note: this does not change the type of the handle_metadata field to
match the channel transport type. This can be done in the future, but
it requires removing dependencies on the the message class in LLCPP
classes, which is better left to another CL.

Bug: 85734
Change-Id: I63becd8562814cb3c743215bc9437b6f59af1cdc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/602564
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Ian McKellar <ianloic@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
API-Review: Ian McKellar <ianloic@google.com>