[fxfs] Fix assertion that could fire during shutdown
The corrupt tests would occasionally show evidence of a panic (without
actually failing), due to reservation amounts getting out-of-sync when
futures got dropped at shutdown.

This CL fixes it so that futures should no longer get dropped in the
middle of a transaction.  This involved fixing VFS so that connections
are only terminated between processing requests, and fixing an issue
with the journal flush task whereby it would drop the compaction task if
it encountered a flush error (which the corruption tests would trigger).

Whilst debugging this, I took the opportunity to refactor Reservation so
that it supports sub-reservations (rather than holds) and supports
reservations where the owner is something other than Arc<dyn Allocator>
(it's now possible for it to be a reference).  I renamed the methods
that require some care to "forget" to match Rust's function of a similar
name.

Change-Id: Ib32f32a2e44de9c69ef3815bdf0099f3ffa1e660
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/609107
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>