[storage] Allow connections to services on R/O filesystems
Connections to services are usually made with read and write rights
(fdio_service_connect does this), but these would always be rejected if
a filesystem was marked as read-only.  This change allows connection to
services even if the filesystem is read-only.  The parent connection
rights should still be honored, so write access would still be required
to the parent connection.

Bug: 89869
Change-Id: Icbc804e0dbb592402e3a4b8453fd7ef96952865f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/617241
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>
Reviewed-by: Brett Wilson <brettw@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>