[Rust VFS] Allow seeking past end of file
Not allowing a zero position seek of a zero capacity file prevents
opening empty files, which broke the rust bootfs.

As Chris pointed out, per the standard we should actually allow seeking
an arbitrary amount past the end of the file. We're also meant to fill
the gap with zeroes if we ever write past the end of the file, and I'll
check that we do that correctly separately.

Documentation:
https://pubs.opengroup.org/onlinepubs/009696799/functions/lseek.html

Tested:
*Unit tests
*Created an empty file in '/boot', tried "cat david.txt" before and
 after

Change-Id: I761ebaa0a3f797acfa6ed0f11f5ed3303a72d7e1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/633110
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: David Hastings <dahastin@google.com>