[vfs] Support handling vmo dirty requests in PagedVfs
This CL allows PagedVfs to control the transition of the VMO page state
from clean to dirty. It introduce a method, PagedVfs::VmoDirty(), in
which filesystem can receive requests for the transition.
If the transition is acceptable, filesystem can handle it by doing some
specific works.
In response to the request, it should request ZX_PAGER_OP_DIRTY so that
kernel transit the pages to dirty and wake up waiters.

Test: fx test fs-connection-test
Change-Id: I00957b1d42dcde2d2714d8c655b6103c36e35e4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/645261
Reviewed-by: Chris Tam <godtamit@google.com>
Commit-Queue: Chris Tam <godtamit@google.com>