[storage][fdf] Provide read/write replacement
Provides a drop-in replacement for calling read
or write on a file descriptor.  The file descriptor
must refer to a fuchsia_io Node, which must be backed
by a block device.

This helper function will be used to replace the instances
of read and write in the storage stack.

Test: A test has been added in the fvm_integration_test.
Change-Id: Ia24c51679907d48de09ad5e690373694af7b1f70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/648110
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: Garratt Gallagher <garratt@google.com>