[fs_management] separate out options, add unit tests
Move all the options to their own header file, add the ability to
generate the cli options based on the option structs, and add unit tests
for that conversion to make sure it's right.

This adds a natural place for doing the conversion to the startup fidl
struct when we add component launching, as well as a way to test that
conversion.

Change-Id: I9bc0d561e821ad9d99733176b8e1459e92ecb4df
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/658507
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Stephen Demos <sdemos@google.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>